refactor: Organizes the src/core folder
