+++ superset-frontend/src/core/index.ts
@@ -16,9 +16,32 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+import { core as coreType } from '@apache-superset/core';
+import { getExtensionsContextValue } from '../extensions/ExtensionsContextUtils';
+import { Disposable } from './models';
+
+export const registerViewProvider: typeof coreType.registerViewProvider = (
+  id,
+  viewProvider,
+) => {
+  const { registerViewProvider: register, unregisterViewProvider: unregister } =
+    getExtensionsContextValue();
+  register(id, viewProvider);
+  return new Disposable(() => unregister(id));
+};
+
+const { GenericDataType } = coreType;
+
+export const core: typeof coreType = {
+  GenericDataType,
+  registerViewProvider,
+  Disposable,
+};
+
 export * from './authentication';
-export * from './core';
 export * from './commands';
 export * from './extensions';
 export * from './environment';
+export * from './models';
 export * from './sqlLab';
+export * from './utils';
